@@ -5,16 +5,23 @@
   "version": "3.1.0",
   "scripts": {
     "start": "razzle start",
-    "prebuild": "./scripts/prebuild-fix-modules.sh",
+    "postinstall": "yarn omelette && yarn patches",
+    "omelette": "ln -sf node_modules/@plone/volto/ omelette",
+    "patches": "/bin/bash patches/patchit.sh > /dev/null 2>&1 ||true",
     "build": "razzle build",
     "test": "razzle test --env=jsdom --passWithNoTests",
     "start:prod": "NODE_ENV=production node build/server.js",
     "i18n": "NODE_ENV=production node node_modules/@plone/volto/src/i18n.js",
-    "develop": "missdev --config=jsconfig.json --output=develop"
+    "develop": "missdev --config=jsconfig.json --output=addons"
   },
+  "private": true,
+  "workspaces": [
+    "src/addons/*"
+  ],
+  "addons": [],
   "jest": {
-    "snapshotSerializers": [
-      "enzyme-to-json/serializer"
+    "modulePathIgnorePatterns": [
+      "api"
     ],
     "transform": {
       "^.+\\.js(x)?$": "babel-jest",
@@ -28,7 +35,9 @@
       "/node_modules/(?!@plone/volto).+\\.js$"
     ],
     "moduleNameMapper": {
+      "@plone/volto/babel": "<rootDir>/node_modules/@plone/volto/babel",
       "@plone/volto/(.*)$": "<rootDir>/node_modules/@plone/volto/src/$1",
+      "load-volto-addons": "<rootDir>/node_modules/@plone/volto/jest-addons-loader.js",
       "@package/(.*)$": "<rootDir>/src/$1",
       "~/(.*)$": "<rootDir>/src/$1"
     },
@@ -42,21 +51,48 @@
     },
     "setupFiles": [
       "@plone/volto/test-setup.js"
-    ]
+    ],
+    "globals": {
+      "__DEV__": true
+    }
   },
   "prettier": {
     "trailingComma": "all",
-    "singleQuote": true
+    "singleQuote": true,
+    "overrides": [
+      {
+        "files": "*.overrides",
+        "options": {
+          "parser": "less"
+        }
+      }
+    ]
   },
   "stylelint": {
     "extends": [
-      "stylelint-config-standard",
-      "stylelint-config-idiomatic-order",
-      "./node_modules/prettier-stylelint/config.js"
-    ]
+      "stylelint-config-idiomatic-order"
+    ],
+    "plugins": [
+      "stylelint-prettier"
+    ],
+    "rules": {
+      "prettier/prettier": true,
+      "rule-empty-line-before": [
+        "always-multi-line",
+        {
+          "except": [
+            "first-nested"
+          ],
+          "ignore": [
+            "after-comment"
+          ]
+        }
+      ]
+    },
+    "ignoreFiles": "theme/themes/default/**/*.overrides"
   },
   "engines": {
-    "node": "^8 || ^10"
+    "node": "^10 || ^12 || ^14"
   },
   "dependencies": {
     "@ckeditor/ckeditor5-basic-styles": "^11.1.4",
@@ -68,15 +104,17 @@
     "@ckeditor/ckeditor5-paragraph": "^11.0.5",
     "@ckeditor/ckeditor5-react": "^1.1.3",
     "@n8tb1t/use-scroll-position": "^1.0.43",
-    "@plone/volto": "github:tiberiuichim/volto#kitchen_sink",
+    "@plone/volto": "12.2.0",
     "compression-webpack-plugin": "3.1.0",
     "draft-js-alignment-plugin": "^2.0.6",
     "draftjs-to-html": "^0.8.4",
     "mgr-pdf-viewer-react": "^1.0.3",
+    "react-component-queries": "2.3.0",
     "react-loadable": "^5.5.0",
     "react-scrollspy": "^3.4.2",
     "react-sizeme": "2.6.12",
-    "react-visibility-sensor": "^5.1.1"
+    "react-visibility-sensor": "^5.1.1",
+    "volto-slate": "^2.1.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.8.4",
@@ -108,22 +146,7 @@
     ">1%",
     "last 4 versions",
     "Firefox ESR",
-    "not ie < 9"
-  ],
-  "resolutions": {
-    "browserslist": "4.8.3"
-  },
-  "customizationPaths": [
-    "src/customizations/volto/"
-  ],
-  "optimizeLiftPackages": [
-    "immutable"
-  ],
-  "manuallyTranspile": [
-    "prepend-http",
-    "normalize-url",
-    "cidr-regex",
-    "url-regex",
-    "lodash-es"
+    "not ie 11",
+    "not dead"
   ]
 }
