@@ -7,21 +7,28 @@ import React, { Component, Fragment } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { compose } from 'redux';
-import { asyncConnect } from 'redux-connect';
-import { Segment } from 'semantic-ui-react';
-import Raven from 'raven-js';
+import { asyncConnect } from '@plone/volto/helpers';
+import { Segment, Grid } from 'semantic-ui-react';
 import { renderRoutes } from 'react-router-config';
 import { Slide, ToastContainer, toast } from 'react-toastify';
 import split from 'lodash/split';
 import join from 'lodash/join';
 import trim from 'lodash/trim';
 import cx from 'classnames';
-import { Grid } from 'semantic-ui-react';
+import config from '@plone/volto/registry';
 
 import Error from '@plone/volto/error';
 
-import { Footer, Header, Icon, Messages } from '@plone/volto/components';
-import { BodyClass, getBaseUrl, getView } from '@plone/volto/helpers';
+import {
+  Breadcrumbs,
+  Footer,
+  Header,
+  Icon,
+  OutdatedBrowser,
+  AppExtras,
+  SkipLinks,
+} from '@plone/volto/components';
+import { BodyClass, getBaseUrl, getView, isCmsUi } from '@plone/volto/helpers';
 import {
   getBreadcrumbs,
   getContent,
@@ -31,10 +38,13 @@ import {
   purgeMessages,
 } from '@plone/volto/actions';
 
+import clearSVG from '@plone/volto/icons/clear.svg';
+import MultilingualRedirector from '@plone/volto/components/theme/MultilingualRedirector/MultilingualRedirector';
+
+import * as Sentry from '@sentry/browser';
+
 import PageHeader from '~/components/theme/Header/PageHeader';
 // import PageHeaderBg from '~/components/theme/Header/PageHeaderBg';
-
-import clearSVG from '@plone/volto/icons/clear.svg';
 
 /**
  * @export
@@ -60,15 +70,18 @@ class App extends Component {
   };
 
   /**
-   * ComponentDidMount
-   * @method ComponentDidMount
+   * ComponentDidCatch
+   * @method ComponentDidCatch
    * @param {string} error  The error
    * @param {string} info The info
    * @returns {undefined}
    */
-  componentDidMount() {
-    if (__CLIENT__ && process.env.SENTRY_DSN) {
-      Raven.config(process.env.SENTRY_DSN).install();
+  componentDidCatch(error, info) {
+    this.setState({ hasError: true, error, errorInfo: info });
+    if (__CLIENT__) {
+      if (window?.env?.RAZZLE_SENTRY_DSN || __SENTRY__?.SENTRY_DSN) {
+        Sentry.captureException(error);
+      }
     }
   }
 
@@ -77,7 +90,7 @@ class App extends Component {
    * @param {Object} nextProps Next properties
    * @returns {undefined}
    */
-  componentWillReceiveProps(nextProps) {
+  UNSAFE_componentWillReceiveProps(nextProps) {
     if (nextProps.pathname !== this.props.pathname) {
       this.props.purgeMessages();
 
@@ -89,7 +102,7 @@ class App extends Component {
 
   checkProps(props) {
     const hideMenuConditions = ['add', 'edit', 'contents'];
-    const pathname = props.pathname && props.pathname.split('/');
+    const pathname = props.pathname?.split('/');
     const pageType = pathname[pathname.length - 1];
     if (hideMenuConditions.includes(pageType)) {
       return true;
@@ -98,33 +111,29 @@ class App extends Component {
   }
 
   /**
-   * ComponentDidCatch
-   * @method ComponentDidCatch
-   * @param {string} error  The error
-   * @param {string} info The info
-   * @returns {undefined}
-   */
-  componentDidCatch(error, info) {
-    this.setState({ hasError: true, error, errorInfo: info });
-    if (__CLIENT__ && process.env.SENTRY_DSN) {
-      Raven.captureException(error, { extra: info });
-    }
-  }
-
-  /**
    * Render method.
    * @method render
    * @returns {string} Markup for the component.
    */
   render() {
-    //
+    const { views } = config;
     const path = getBaseUrl(this.props.pathname);
     const action = getView(this.props.pathname);
+    const isCmsUI = isCmsUi(this.props.pathname);
     const hideMenu = this.checkProps(this.props);
+    const ConnectionRefusedView = views.errorViews.ECONNREFUSED;
     return (
       <Fragment>
         <BodyClass className={`view-${action}view`} />
         {/* <BodyClass className={this.props.pathname === '/' ? 'homepage' : ''} /> */}
+        {/* Body class depending on content type */}
+        {this.props.content && this.props.content['@type'] && (
+          <BodyClass
+            className={`contenttype-${this.props.content['@type']
+              .replace(' ', '-')
+              .toLowerCase()}`}
+          />
+        )}
 
         {/* Body class depending on sections */}
         <BodyClass
@@ -132,6 +141,10 @@ class App extends Component {
             [trim(join(split(this.props.pathname, '/'), ' section-'))]:
               this.props.pathname !== '/',
             siteroot: this.props.pathname === '/',
+            'is-authenticated': !!this.props.token,
+            'is-anonymous': !this.props.token,
+            'cms-ui': isCmsUI,
+            'public-ui': !isCmsUI,
           })}
         />
         {this.props.pathname === '/' ? (
@@ -142,19 +155,25 @@ class App extends Component {
               pathname={path}
               navigationItems={this.props.navigation}
             />
-            <Segment className="content-area">
-              <main>
-                <Messages />
-                {this.state.hasError ? (
-                  <Error
-                    message={this.state.error.message}
-                    stackTrace={this.state.errorInfo.componentStack}
-                  />
-                ) : (
-                  renderRoutes(this.props.route.routes)
-                )}
-              </main>
-            </Segment>
+            <MultilingualRedirector pathname={this.props.pathname}>
+              <Segment basic className="content-area">
+                <main>
+                  <OutdatedBrowser />
+                  {this.props.connectionRefused ? (
+                    <ConnectionRefusedView />
+                  ) : this.state.hasError ? (
+                    <Error
+                      message={this.state.error.message}
+                      stackTrace={this.state.errorInfo.componentStack}
+                    />
+                  ) : (
+                    renderRoutes(this.props.route.routes, {
+                      staticContext: this.props.staticContext,
+                    })
+                  )}
+                </main>
+              </Segment>
+            </MultilingualRedirector>
           </React.Fragment>
         ) : (
           <div className="content-page">
@@ -184,7 +203,6 @@ class App extends Component {
                   largeScreen={hideMenu ? 12 : 6}
                 >
                   <main className="content-page">
-                    <Messages />
                     <div className="editor-toolbar-wrapper" />
                     {this.state.hasError ? (
                       <Error
@@ -222,13 +240,20 @@ class App extends Component {
           }
         />
         <Footer />
+        <AppExtras {...this.props} />
       </Fragment>
     );
   }
 }
 
 export const __test__ = connect(
-  (state, props) => ({ pathname: props.location.pathname }),
+  (state, props) => ({
+    pathname: props.location.pathname,
+    token: state.userSession.token,
+    content: state.content.data,
+    apiError: state.apierror.error,
+    connectionRefused: state.apierror.connectionRefused,
+  }),
   { purgeMessages },
 )(App);
 
@@ -247,7 +272,13 @@ export default compose(
     {
       key: 'navigation',
       promise: ({ location, store: { dispatch } }) =>
-        __SERVER__ && dispatch(getNavigation(getBaseUrl(location.pathname), 4)),
+        __SERVER__ &&
+        dispatch(
+          getNavigation(
+            getBaseUrl(location.pathname),
+            config.settings.navDepth,
+          ),
+        ),
     },
     {
       key: 'types',
@@ -262,7 +293,11 @@ export default compose(
   ]),
   connect(
     (state, props) => ({
-      pathname: state.router.location.pathname, //props.location.pathname,
+      pathname: state.router.location.pathname,
+      token: state.userSession.token,
+      content: state.content.data,
+      apiError: state.apierror.error,
+      connectionRefused: state.apierror.connectionRefused, //props.location.pathname,
     }),
     { purgeMessages },
   ),
