@@ -5,17 +5,21 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { isMatch } from 'lodash';
 import { compose } from 'redux';
+import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import { defineMessages, injectIntl, intlShape } from 'react-intl';
 import cx from 'classnames';
+import { getBaseUrl } from '@plone/volto/helpers';
 import { BodyClass } from '@plone/volto/helpers';
 import rightKey from '@plone/volto/icons/right-key.svg';
 import backIcon from '@plone/volto/icons/back.svg';
 import { Icon } from '@plone/volto/components';
-import { settings } from '~/config';
+import config from '@plone/volto/registry';
 import { getBasePath } from '~/helpers';
 
-import bgimage from './home.jpg';
+import { getNavigation } from '@plone/volto/actions';
+
+import bgimage from '~/components/theme/Navigation/home.jpg';
 
 const messages = defineMessages({
   closeMobileMenu: {
@@ -31,19 +35,21 @@ const messages = defineMessages({
 function getPath(url) {
   if (!url) return '';
   return url
-    .replace(settings.apiPath, '')
-    .replace(settings.internalApiPath, '');
+    .replace(config.settings.apiPath, '')
+    .replace(config.settings.internalApiPath, '');
 }
 
 class Navigation extends Component {
   static propTypes = {
     pathname: PropTypes.string.isRequired,
-    // items: PropTypes.arrayOf(
-    //   PropTypes.shape({
-    //     title: PropTypes.string,
-    //     url: PropTypes.string,
-    //   }),
-    // ).isRequired,
+    getNavigation: PropTypes.func.isRequired,
+    items: PropTypes.arrayOf(
+      PropTypes.shape({
+        title: PropTypes.string,
+        url: PropTypes.string,
+      }),
+    ).isRequired,
+    lang: PropTypes.string.isRequired,
   };
 
   constructor(props) {
@@ -68,13 +74,29 @@ class Navigation extends Component {
   }
 
   componentDidMount() {
+    const { settings } = config;
+    this.props.getNavigation(
+      getBaseUrl(this.props.pathname),
+      settings.navDepth,
+    );
     document.addEventListener('resize', this.isMobile);
     this.isMobile();
   }
 
-  componentWillReceiveProps(nextProps) {
+  /**
+   * Component will receive props
+   * @method componentWillReceiveProps
+   * @param {Object} nextProps Next properties
+   * @returns {undefined}
+   */
+  UNSAFE_componentWillReceiveProps(nextProps) {
+    const { settings } = config;
     if (nextProps.pathname !== this.props.pathname) {
       this.closeMobileMenu();
+      this.props.getNavigation(
+        getBaseUrl(nextProps.pathname),
+        settings.navDepth,
+      );
     }
   }
 
@@ -153,7 +175,7 @@ class Navigation extends Component {
 
   render() {
     // console.log('------------', this.props.navigation);
-    const navigation = this.formatNavUrl(this.props.navigation.items);
+    const navigation = this.formatNavUrl(this.props.items);
 
     return (
       <div id="app-menu-content">
@@ -168,7 +190,7 @@ class Navigation extends Component {
         <div
           id="menu-background"
           className="background-img"
-          style={{ backgroundImage: `url(${bgimage})` }}
+          style={{ backgroundImage: `url(${bgimage})`, height: '500%' }}
         />
         <div
           id="menu-background-overlay"
@@ -217,9 +239,9 @@ class Navigation extends Component {
           <div className="first-level">
             {navigation.map((item, index) => (
               <div
-                key={getPath(item['@id'])}
+                key={getPath(item.url)}
                 className={
-                  this.isActive(getPath(item['@id']))
+                  this.isActive(getPath(item.url))
                     ? 'menu-item active'
                     : 'menu-item'
                 }
@@ -241,14 +263,15 @@ class Navigation extends Component {
                     {item.title}
                   </a>
                 ) : (
-                  <Link to={getPath(item['@id'])} key={getPath(item['@id'])}>
+                  <Link to={getPath(item.url)} key={getPath(item.url)}>
                     {item.title}
                   </Link>
                 )}
               </div>
             ))}
           </div>
-          {this.state.subMenu.items && this.state.subMenu.items.length ? (
+          {console.log(this.state.subMenu.items)}
+          {this.state.subMenu.items && this.state.subMenu.items.length > 0 ? (
             <div className="second-level">
               <Icon
                 name={backIcon}
@@ -289,10 +312,10 @@ class Navigation extends Component {
                 </div>
               ))}
             </div>
-          ) : (
-            <div />
-          )}
-          {this.state.subTopics.items && this.state.subTopics.items.length ? (
+          ) : null}
+          {console.log('sunto', this.state.subTopics.items)}
+          {this.state.subTopics.items &&
+          this.state.subTopics.items.length > 0 ? (
             <div className="third-level">
               <Icon
                 name={backIcon}
@@ -327,24 +350,20 @@ class Navigation extends Component {
                 </div>
               ))}
             </div>
-          ) : (
-            ''
-          )}
+          ) : null}
         </div>
       </div>
     );
   }
 }
-// export default compose(
-//   injectIntl,
-//   connect(
-//     state => ({
-//       items:
-//         state.content.data &&
-//         state.content.data['@components'].navigation.items,
-//     }),
-//     {},
-//   ),
-// )(Navigation);
 
-export default compose(injectIntl)(Navigation);
+export default compose(
+  injectIntl,
+  connect(
+    state => ({
+      items: state.navigation.items,
+      lang: state.intl.locale,
+    }),
+    { getNavigation },
+  ),
+)(Navigation);
