@@ -3,8 +3,8 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { BodyClass } from '@plone/volto/helpers';
 
-import ecLogo from './ec_white_negative_color.png';
-import eeaLogo from './eea_white.png';
+import ecLogo from '~/components/theme/Header/ec_white_negative_color.png';
+import eeaLogo from '~/components/theme/Header/eea_white.png';
 
 import {
   Logo,
@@ -68,7 +68,6 @@ class Header extends Component {
             </div>
             <Navigation
               isHomepage={this.props.homepage}
-              navigation={this.props.navigationItems}
               pathname={this.props.pathname}
             />
           </React.Fragment>
