@@ -3,7 +3,7 @@
  * @module index
  */
 
-import start from 'volto-base/start-server';
+import start from '@plone/volto/start-server';
 
 const reloadServer = start();
 
