@@ -5,8 +5,7 @@
 
 import { App } from '@plone/volto/components';
 import { defaultRoutes } from '@plone/volto/routes';
-
-import { addonRoutes } from '~/config';
+import config from '@plone/volto/registry';
 
 /**
  * Routes array.
@@ -19,28 +18,7 @@ const routes = [
     component: App, // Change this if you want a different component
     routes: [
       // Add your routes here
-      // {
-      //   path: '/',
-      //   component: HomepageView,
-      //   exact: true,
-      // },
-      // {
-      //   path: '/manage-slider',
-      //   component: EditSlider,
-      // },
-      // {
-      //   path: '*/**/manage-slider',
-      //   component: EditSlider,
-      // },
-
-      // addon routes have a higher priority then default routes
-      // {
-      //   path: '/',
-      //   component: HomepageView,
-      //   exact: true,
-      // },
-      ...(addonRoutes || []),
-
+      ...(config.addonRoutes || []),
       ...defaultRoutes,
     ],
   },
