@@ -1,3 +1,2 @@
 import 'semantic-ui-less/semantic.less';
-// import '@plone/volto/../theme/theme.less';
-import '@plone/volto/../theme/themes/pastanaga/extras/extras.less';
+//import '@plone/volto/../theme/themes/pastanaga/extras/extras.less';
